[new] fully working LAPIC timer
currently in periodic mode, as easier and more common, but if different
timeslices for scheduling is needed, either take multiple of waiting
time or implement one-shot mode

+++ src/kernel/hardware/hpet/hpet.c
@@ -45,7 +45,7 @@ void hpet_init(void)
 }
 
 // wait for us nanoseconds
-void hpet_usleep(size_t us)
+void hpet_usleep(uint64_t us)
 {
     uint32_t clock_period = hpet_regs->capabilities >> 32;
